New "jstables" (JUMP_SLOT tables) diagram for lazily-resolved function calls
This requires an edge between the PLTs to keep Graphviz from using a stupid layout.  Said edge
produces the desired output file, but causes Graphviz to emit an "error" and exit with nonzero
status.  So this patch also introduces a workaround in the Makefile that lets us ignore the
exit status of a dot invocation.